Improve header component
