+++ src/gui/static/src/app/components/pages/send-skycoin/send-skycoin.component.html
@@ -22,7 +22,7 @@
       <textarea formControlName="notes" id="notes" row="2"></textarea>
     </div>
     <app-button #button (action)="send()" class="primary" [disabled]="!form.valid">
-      Create
+      Send
     </app-button>
   </div>
 </div>
