+++ src/gui/static/src/app/services/blockchain.service.ts
@@ -1,13 +1,35 @@
 import { Injectable } from '@angular/core';
 import { ApiService } from './api.service';
 import { Observable } from 'rxjs/Observable';
+import { IntervalObservable } from 'rxjs/observable/IntervalObservable';
+import { Subject } from 'rxjs/Subject';
+import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { TellerConfig } from '../app.datatypes';
+import { WalletService } from './wallet.service';
 
 @Injectable()
 export class BlockchainService {
 
+  private progressSubject: Subject<any> = new BehaviorSubject<any>(null);
+
+  get progress() {
+    return this.progressSubject.asObservable();
+  }
+
   constructor(
     private apiService: ApiService,
-  ) { }
+    private walletService: WalletService,
+  ) {
+    IntervalObservable
+      .create(2000)
+      .flatMap(() => this.getBlockchainProgress())
+      .takeWhile((response: any) => !response.current || response.current !== response.highest)
+      .subscribe(
+        response => this.progressSubject.next(response),
+        error => console.log(error),
+        () => this.completeLoading()
+      );
+  }
 
   addressTransactions(id): Observable<any> {
     return this.apiService.get('explorer/address', { address: id });
@@ -44,8 +66,13 @@ export class BlockchainService {
     return this.blocks(1).map(blocks => blocks[0]);
   }
 
-  progress() {
+  getBlockchainProgress() {
     return this.apiService.get('blockchain/progress');
+  }
+
+  private completeLoading() {
+    this.progressSubject.next({ current: 999999999999, highest: 999999999999 });
+    this.walletService.refreshBalances();
   }
 
   private retrieveInputAddress(input: string) {
